# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "タスみる API",
    "description": "仕様書からタスクを自動生成し、担当候補を提案するAPI",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/projects": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Create Project",
        "description": "プロジェクトを作成する（チームに紐づく）",
        "operationId": "create_project_api_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}": {
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project",
        "operationId": "get_project_api_projects__project_id__get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/members": {
      "put": {
        "tags": [
          "projects"
        ],
        "summary": "Set Project Members",
        "description": "プロジェクトのメンバー情報(Phase 6)を設定する。\n\n`backend.pipeline.members.runner.build_member_directory`（既存の取り込み\nロジック）をそのまま呼ぶだけ。スキルレベル等はここでも一切作り出さない\n——`body.members`に含まれていない情報は現れない。",
        "operationId": "set_project_members_api_projects__project_id__members_put",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMembersRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDirectory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "projects"
        ],
        "summary": "Get Project Members",
        "operationId": "get_project_members_api_projects__project_id__members_get",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemberDirectory"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/generate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Generate Project",
        "description": "Phase 3-9パイプライン全体を実行するジョブを開始する。\n\n**この関数はパイプラインを実行しない。** ジョブレコードを作成して\nバックグラウンドタスクとしてスケジュールし、即座にjob_idを返す\n（依頼の\"Avoid making the frontend wait indefinitely for one\nsynchronous HTTP request\"に対応）。",
        "operationId": "generate_project_api_projects__project_id__generate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Availability": {
        "properties": {
          "available_hours_per_week": {
            "type": "number",
            "title": "Available Hours Per Week"
          },
          "working_days": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Working Days"
          },
          "current_assigned_hours": {
            "type": "number",
            "title": "Current Assigned Hours",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "available_hours_per_week"
        ],
        "title": "Availability",
        "description": "稼働可能時間・稼働日・現在の割当状況"
      },
...
      "Constraint": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "value": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "max_hours": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Hours"
          }
        },
        "type": "object",
        "required": [
          "type"
        ],
        "title": "Constraint",
        "description": "構造化された制約条件（自由記述の文字列ではない）\n\ntypeに応じて意味のあるフィールドが変わる:\n  - scope_restriction  : value に対象範囲（例: \"frontend\"）\n  - day_unavailable    : value に曜日名（例: \"Monday\"）\n  - max_hours_per_week : max_hours に週あたりの上限時間\n  - requires_review    : value にレビュー要件の説明（例: \"senior member\"）"
      },
...
      "GenerateRequest": {
        "properties": {
          "document_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Text",
            "description": "指定した場合、生成前にプロジェクトの仕様書本文を更新する"
          },
          "use_assignment_llm_reasoning": {
            "type": "boolean",
            "title": "Use Assignment Llm Reasoning",
            "description": "Phase 7 Step 3（LLMによる補足説明）を有効にするか（既定オフ）",
            "default": false
          },
          "use_duplicate_llm_verification": {
            "type": "boolean",
            "title": "Use Duplicate Llm Verification",
            "description": "Phase 8 CHECK 2の重複候補についてLLM検証を行うか（既定オフ）",
            "default": false
          }
        },
        "type": "object",
        "title": "GenerateRequest"
      },
      "GenerateResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "status": {
            "type": "string",
            "const": "queued",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "status"
        ],
        "title": "GenerateResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Member": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "title": "Skills"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years",
            "description": "全体的な業務経験年数（個々のスキルのexperience_yearsとは別）"
          },
          "availability": {
            "$ref": "#/components/schemas/Availability"
          },
          "constraints": {
            "items": {
              "$ref": "#/components/schemas/Constraint"
            },
            "type": "array",
            "title": "Constraints"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "availability"
        ],
        "title": "Member",
        "description": "タスクアサインの対象となる、構造化されたメンバー情報"
      },
      "MemberDirectory": {
        "properties": {
          "team_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/Member"
            },
            "type": "array",
            "title": "Members"
          },
          "issues": {
            "items": {
              "$ref": "#/components/schemas/backend__pipeline__members__schema__ValidationIssue"
            },
            "type": "array",
            "title": "Issues"
          },
          "updated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updated At",
            "description": "ISO8601形式の更新時刻(UTC)"
          }
        },
        "type": "object",
        "required": [
          "members"
        ],
        "title": "MemberDirectory",
        "description": "members.json の全体（1チーム分のメンバー一覧）"
      },
...
      "ProjectCreateRequest": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name",
            "description": "プロジェクト名（表示用、任意）"
          },
          "document_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Text",
            "description": "仕様書本文（後から/generate時に渡すことも可能）"
          }
        },
        "type": "object",
        "title": "ProjectCreateRequest"
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "team_id": {
            "type": "string",
            "title": "Team Id"
          },
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "has_document": {
            "type": "boolean",
            "title": "Has Document"
          },
          "has_members": {
            "type": "boolean",
            "title": "Has Members"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "team_id",
          "has_document",
          "has_members",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectResponse"
      },
...
      "SetMembersRequest": {
        "properties": {
          "members": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Members",
            "description": "Phase 6のMemberスキーマに沿った生レコードの配列。ここでは一切値を作り出さない。"
          }
        },
        "type": "object",
        "required": [
          "members"
        ],
        "title": "SetMembersRequest"
      },
...
      "Skill": {
        "properties": {
          "skill": {
            "type": "string",
            "title": "Skill"
          },
          "level": {
            "type": "integer",
            "title": "Level"
          },
          "experience_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Experience Years"
          }
        },
        "type": "object",
        "required": [
          "skill",
          "level"
        ],
        "title": "Skill",
        "description": "1つのスキルとその習熟度"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "backend__pipeline__members__schema__ValidationIssue": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "member_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Id"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ValidationIssue",
        "description": "Member Validationステージが検出した問題点"
      },
...
    }
  }
}